reforms ang page components
