+++ src/js/pages/Ang/components/Layout.tsx
@@ -1,22 +1,24 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import { pageView, errorEvent, ACTIONS } from '../../util/analytics';
-import GenericError, { BalpreetSingh } from '../../components/GenericError';
-import { SOURCES, TEXTS } from '../../constants';
-import ShabadContent from '../../components/ShabadContent';
-import { toAngURL } from '../../util';
-import BreadCrumb from '../../components/Breadcrumb';
+
+import GenericError, { BalpreetSingh } from '@/components/GenericError';
+import ShabadContent from '@/components/ShabadContent';
+import BreadCrumb from '@/components/Breadcrumb';
+import { SOURCES, TEXTS } from '@/constants';
+import { pageView, errorEvent, ACTIONS, toAngURL } from '@/util';
 
 export const Stub = () => <div className="spinner" />;
 
-export default class Layout extends React.PureComponent {
-  static propTypes = {
-    ang: PropTypes.number,
-    source: PropTypes.oneOf(Object.keys(SOURCES)),
-    highlight: PropTypes.number,
-    data: PropTypes.object.isRequired,
-  };
+type Sources = keyof typeof SOURCES;
+interface ILayoutProps {
+  ang: number
+  source: Sources
+  highlight: number
+  data: any
+}
+
+export default class Layout extends React.PureComponent<ILayoutProps> {
+
   render() {
     const { ang, source, highlight, data } = this.props;
